feat: Remove duplicated file upload logic
