+++ editor.planx.uk/src/@planx/components/DrawBoundary/Public/index.tsx
@@ -9,16 +9,15 @@ import {
   MapFooter,
 } from "@planx/components/shared/Preview/MapContainer";
 import QuestionHeader from "@planx/components/shared/Preview/QuestionHeader";
+import { PrivateFileUpload } from "@planx/components/shared/PrivateFileUpload/PrivateFileUpload";
 import type { PublicProps } from "@planx/components/ui";
 import type { Geometry } from "@turf/helpers";
 import { Store, useStore } from "pages/FlowEditor/lib/store";
 import React, { useEffect, useRef, useState } from "react";
 
 import { DrawBoundary, PASSPORT_UPLOAD_KEY } from "../model";
-import Upload from "./Upload";
 
 export type Props = PublicProps<DrawBoundary>;
-export type SelectedFile = FileUploadSlot;
 
 export default function Component(props: Props) {
   const isMounted = useRef(false);
@@ -27,19 +26,17 @@ export default function Component(props: Props) {
   const previousArea =
     props.previouslySubmittedData?.data?.[props.dataFieldArea];
   const previousFile =
-    props.previouslySubmittedData?.data?.[PASSPORT_UPLOAD_KEY]?.[0];
+    props.previouslySubmittedData?.data?.[PASSPORT_UPLOAD_KEY];
   const startPage = previousFile ? "upload" : "draw";
   const [page, setPage] = useState<"draw" | "upload">(startPage);
   const passport = useStore((state) => state.computePassport());
   const [boundary, setBoundary] = useState<Boundary>(previousBoundary);
-  const [selectedFile, setSelectedFile] = useState<SelectedFile | undefined>(
-    previousFile
-  );
+  const [slots, setSlots] = useState<FileUploadSlot[]>(previousFile ?? []);
   const [area, setArea] = useState<number | undefined>(previousArea);
   const environment = useStore((state) => state.previewEnvironment);
 
   useEffect(() => {
-    if (isMounted.current) setSelectedFile(undefined);
+    if (isMounted.current) setSlots([]);
     isMounted.current = true;
 
     const areaChangeHandler = ({ detail }: { detail: string }) => {
@@ -67,14 +64,12 @@ export default function Component(props: Props) {
       map?.removeEventListener("areaChange", areaChangeHandler);
       map?.removeEventListener("geojsonChange", geojsonChangeHandler);
     };
-  }, [page, setArea, setBoundary, setSelectedFile]);
+  }, [page, setArea, setBoundary, setSlots]);
 
   return (
     <Card
       handleSubmit={handleSubmit}
-      isValid={
-        props.hideFileUpload ? true : Boolean(boundary || selectedFile?.url)
-      }
+      isValid={props.hideFileUpload ? true : Boolean(boundary || slots[0]?.url)}
     >
       {getBody()}
     </Card>
@@ -153,12 +148,12 @@ export default function Component(props: Props) {
             howMeasured={props.howMeasured}
             definitionImg={props.definitionImg}
           />
-          <Upload setFile={setSelectedFile} initialFile={selectedFile} />
+          <PrivateFileUpload slots={slots} setSlots={setSlots} maxFiles={1} />
           <Box sx={{ textAlign: "right" }}>
             <Link
               component="button"
               onClick={() => setPage("draw")}
-              disabled={Boolean(selectedFile?.url)}
+              disabled={Boolean(slots[0]?.url)}
             >
               <Typography variant="body2">
                 Draw the boundary on a map instead
@@ -182,7 +177,7 @@ export default function Component(props: Props) {
           boundary && area && props.dataFieldBoundary
             ? area / 10000
             : undefined,
-        [PASSPORT_UPLOAD_KEY]: selectedFile ? [selectedFile] : undefined,
+        [PASSPORT_UPLOAD_KEY]: slots,
       };
     })();
